% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcCellmap.R
\name{abcCellmap}
\alias{abcCellmap}
\title{Human Blood Cell Landscape mapping}
\usage{
abcCellmap(queryData)
}
\arguments{
\item{queryData}{a space-delimited txt format file containing the expression matrix. Each row should be a gene symbol and each column should be a single cell. We prefer the expression data is presented by UMIs (Unique Molecular Identifiers) per gene in each single cell.}
}
\value{
Users can predict the cell types of hematopoietic cells by implementing two approaches (Scmap and Seurat). Cells in our ABC are labeled by 43 different RNA clusters according to unsupervised clustering of single-cell transcriptional profiles, and also labeled by 32 immunophenotypic cell types, involving HSPC, B cell, T cell, NK cell, Neutrophil, Monocyte and Erythrocyte population. The format of result are as follows:

queryCell, Seurat.RNACluster, Seurat.RNACluster.score, Seurat.Immunophenotype, Seurat.Immunophenotype.score, scmap.RNACluster, scmap.RNACluster.score, scmap.Immunophenotype, scmap.Immunophenotype.score, scmap.Cell, scmap.Cell.score, pertype.

Of which, "queryCell" is the cell information in the query data,

"Seurat.RNACluster" is the RNA cluster predicted by Seurat,

"Seurat.RNACluster.score" is the prediction score of RNA cluster by Seurat,

"Seurat.Immunophenotype" is the immunophenotypic cell type predicted by Seurat,

"Seurat.Immunophenotype.score" is the prediction score of immunophenotypic cell type by Seurat,

"scmap.RNACluster" is the RNA cluster predicted by scmap,

"scmap.RNACluster.score" is the prediction score of RNA cluster by scmap,

"scmap.Immunophenotype" is the immunophenotypic cell type predicted by scmap,

"scmap.Immunophenotype.score" is the prediction score of immunophenotypic cell type by scmap,

"scmap.Cell" is the nearest single cell in our ABC reference predicted by scmap,

"scmap.Cell.score" is the prediction score of the nearest cell by scmap,

"pertype" means the percentages of top 2 immunophenotypic cell types in corresponding Seurat.RNACluster result.
}
\description{
Human Blood Cell Landscape R package for single cell mapping.
}
\examples{
mapresult<-abcCellmap(queryData=query.exp)
}
